+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Set up a site language for the fake frontend in tests (#1139)
 
 ## 4.0.3
 
@@ -38,7 +39,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Remove the favorites list dummy placeholder (#1100)
 
 ### Fixed
-- Construct the FE controller differently in V10 (#1125)
+- Construct the FE controller in differently in V10 (#1125)
 - Add localized label for event description in the CSV export (#1115)
 - Stop using the deprecated `BaseScriptClass` (#1111)
 - Stop using the language labels of the lang extension (#1097)